bigdata/producer/requirements.txt
confluent_kafka
requests
pandas